Make ParakeetModelState Equatable and delete its mirror enums
ParakeetModelState was not Equatable, so three surfaces kept byte-identical
mirror enums with hand-written 16-line converters, and STTRouter carried a
manual modelStatesEqual. Move the enum to its own deps-free file (so the
fast-test runner can compile it alongside the policy files), synthesize
Equatable, and:

- delete FirstRunLocalModelState (FirstRunExperience.swift) and its converter
  in PermissionsOnboardingView; call sites pass sttRouter.modelDownloadState
  directly
- delete MeetingWarmupDictationState and its converter in
  MeetingSessionController; MeetingWarmupStatusPolicy takes ParakeetModelState
  (MeetingWarmupMeetingState stays: it bridges TranscriptedCore's
  DiarizationModelState into a runner-compilable policy file)
- delete DictationWarmupPresentationPolicy.ModelState and the
  warmupModelState converter in DictationSessionController
- delete STTRouter.modelStatesEqual in favor of ==
